+++ CHANGELOG.md
@@ -1,5 +1,8 @@
 # Changelog
 
+## Unreleased
+* Add special handling of `<=python` version for OpenSSL compatibility between Julia and Python.
+
 ## v0.1.16 (2025-02-18)
 * Adds file-locking to protect multiple concurrent resolves.